pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "viennatalksbout"
version = "0.1.0"
description = "Real-time tag cloud showing trending topics in Vienna"
requires-python = ">=3.10"
dependencies = [
    "Mastodon.py>=2.1.4",
    "anthropic>=0.40.0",
    # beautifulsoup4 uses Python's built-in html.parser — no extra parser dep needed
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
include = ["viennatalksbout*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: tests that make real HTTP calls to external services",
]

[tool.coverage.run]
source = ["viennatalksbout"]

[tool.coverage.report]
fail_under = 90
show_missing = true
